Fixed bugs, and replaced Duration with just the milliseconds long value (because Kotlinx serialization does not handle Duration properly)

+++ server/src/main/kotlin/data/values/Achievements.kt
@@ -26,7 +26,7 @@ val achievements: List<Achievement>
             "Get any item duration to $infiniteAnimationThreshold or below",
             "NOT THE EYES!",
             listOf(
-                AchievementCondition.anyItem { fillRate { leq(infiniteAnimationThreshold.inWholeMilliseconds.toDouble()) } },
+                AchievementCondition.anyItem { fillRate { leq(infiniteAnimationThreshold.inWholeMilliseconds) } },
             ),
         ),
         Achievement(
